+++ packages/affiliates/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @happyvertical/smrt-affiliates
 
+## 0.19.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.72
+  - @happyvertical/smrt-ads@0.19.72
+  - @happyvertical/smrt-commerce@0.19.72
+  - @happyvertical/smrt-profiles@0.19.72
+  - @happyvertical/smrt-properties@0.19.72
+
+## 0.19.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.71
+  - @happyvertical/smrt-ads@0.19.71
+  - @happyvertical/smrt-commerce@0.19.71
+  - @happyvertical/smrt-profiles@0.19.71
+  - @happyvertical/smrt-properties@0.19.71
+
 ## 0.19.70
 
 ### Patch Changes