+++ packages/properties/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @happyvertical/smrt-properties
 
+## 0.19.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.72
+  - @happyvertical/smrt-profiles@0.19.72
+  - @happyvertical/smrt-projects@0.19.72
+  - @happyvertical/smrt-tenancy@0.19.72
+
+## 0.19.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.71
+  - @happyvertical/smrt-profiles@0.19.71
+  - @happyvertical/smrt-projects@0.19.71
+  - @happyvertical/smrt-tenancy@0.19.71
+
 ## 0.19.70
 
 ### Patch Changes