+++ packages/template-sveltekit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @happyvertical/smrt-template-sveltekit
 
+## 0.19.72
+
+## 0.19.71
+
 ## 0.19.70
 
 ## 0.19.69