+++ packages/voice/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @happyvertical/smrt-voice
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.72
+  - @happyvertical/smrt-assets@0.19.72
+  - @happyvertical/smrt-content@0.19.72
+  - @happyvertical/smrt-tenancy@0.19.72
+  - @happyvertical/smrt-config@0.19.72
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.19.71
+  - @happyvertical/smrt-assets@0.19.71
+  - @happyvertical/smrt-content@0.19.71
+  - @happyvertical/smrt-tenancy@0.19.71
+  - @happyvertical/smrt-config@0.19.71
+
 ## 0.0.18
 
 ### Patch Changes